requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
cryptography==41.0.8
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
